requirements.txt
asgiref==3.4.1
autopep8==1.6.0
Django==4.0.1
django-cors-headers==3.11.0
django-dotenv==1.4.2
djangorestframework==3.13.1
numpy==1.22.1
pandas==1.3.5
psycopg2-binary==2.9.3
pycodestyle==2.8.0
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
